Configure location for the compiled Lambda jar to be uploaded
[Problem]
CodeBuild uploads the S3DropboxLambda jar in a non-deterministic location i.e. the given folder name may be chopped. It
also zips the jar, which is just slightly annoying to deal with.

[Solution]
Instead of using CodeBuild's `OutputArtifacts` feature, upload the artifacts in the post_build stage, specifying the
location via CodeBuild environment variables.

[Test]
Updated the stack with it pointing at a different `S3DropboxLambda` repository branch - the Lambda fat jar as uploaded
to the expected location.