Create CF Stack phone number parameter
[Problem]
A phone number is needed to send messages to the owner of the repo when the build fails. However, since this is personal
information, phone numbers should not be committed to a git repository.

[Solution]
Add a phone number template parameter.

[Test]
Created phone number branches for the S3DropboxLambda and Dossier Latex repositories. The CodeBuild projects were
updated to point at these repositories and ensured that the build both fails and succeeds.  The failure was forced by
introducing an error in one of the Latex projects.