Trigger the Lambda function on PDF package updates
[Problem]
The S3DropboxLambda Lambda function is not triggered when new PDF packages are deployed.

[Solution]
Introducing S3 notification events to trigger the Lambda caused a circulare dependency between the Latex bucket and the
Lambda function.
  * The Latex bucket needs the Lambda ARN for which Lambda function to call to process the S3 event
  * The Lambda needs the Latex bucket for the code location and to create the permission for S3 to invoke the function

The solution was to create a 3 stage deployment:
  1. Deploy the pipelines so that the Lambda function code package was deployed.
  2. Deploy the Lambda function
  3. Deploy the S3 event notification configuration, which points to the Lambda function

Note that the Lambda had to be deployed to a separate bucket because even though the circular dependency was resolved
manually, CloudFormation does not permit the third stage to be deployed if the Lambda function and Latex package exist
in the same bucket.

The sequence of commands starting from the empty state is:
  $ rake dossier:deploy_stage[github_oauth_key, phone_num, account_id, 0]
  $ rake dossier:deploy_stage[github_oauth_key, phone_num, account_id, 1]
  $ rake dossier:deploy_stage[github_oauth_key, phone_num, account_id, 2] // use this for future updates

A stack output containing only the Lambda Pipeline name can be obtained via:
  * `aws cloudformation describe-stacks --stack-name STACK_NAME`
This is so that after the first deployment (0), the pipeline for building and deploying the Lambda code can be watched
so that the user is aware of when the subsequent two deployments can occur (1 and 2). After stage 1 is deployed, stage 2
can be immediately deployed afterwards.

[Test]
Deployed CloudFormation stack starting from an empty state. Triggered the Lambda and the Lambda executed.