Add the encrypted Dropbox token as an environment variable to the Lambda Function
[Problem]
The AWS Lambda function needs the Dropbox token to publish to Dropbox.

[Solution]
Encrypt the Dropbox token using the Lambda KMS key and add the encrypted token as an environment variable in the AWS
Lambda function.

Change the Lambda in the template to be a placeholder function that is updated in the Lambda repository. This decouples
the CloudFormation template and the actual implementation of the AWS Lambda function.

As a side-effect, the CodePipeline had to be moved into the second stage of deployment. The reason for this is because
the Lambda package needs the encrypted value of the Dropbox token.

Using a placeholder Lambda function also permited the S3 notification configuration to be moved to the second deployment
stage.

[Test]
Deploy the CloudFormation template along with the changes that were required in the Lambda Function code.