Update CF to support short term dropbox credentials
## Problem
Dropbox is deprecating long term credentials. The Dossier project needs to be updated to support short term credentials.

## Solution
Add support for the Lambda function having permission to retrieve and store credentials in a serialized object in S3.

Docker is also throttling anonymous access to images and so the Latex Code Build image was moved to ECR. Update
CodeBuild to use the ECR image.

## Test
Updated the CloudFormation in account and tested Dossier build.