Update to the new GitHub authorization & authentication mechanism
## Problem
GitHub has created a new AA mechanism. AWS has integrated with it via CodeStar, which runs as an authentication hook in
the GitHub account. More details on the new GitHub AA can be found here:

* https://github.blog/2021-04-05-behind-githubs-new-authentication-token-formats/

The old mechanism will likely no longer support OAuth in the future, so it is better to be proactive about updating.

## Solution
Update the pipeline to use the new authorization & authentication mechanism using this guide:

* https://docs.aws.amazon.com/codepipeline/latest/userguide/update-github-action-connections.html

## Test
Deployed & manually triggered source release